wip: add `JWT` as possible `AuthorizationMode`
This is a proof-of-concept commit to demonstrate adding the ability to
specify a JWT directly rather than either an access token or private
key.

When a JWT is configured, the Okta `RequestExecutor` class can skip the
JWT generation and go straight to exchanging the JWT for an access
token, which it then appends to the request.

This access token can then be cached normally, as it is in the
`PrivateKey` flow.